Adds shutdown hook and uncaught exception handler.
Added shutdown hook to Main. Running the shutdown method will terminate all thread started by Hydra. If a thread is hung, Hydra will be killed after 30 seconds

Added an uncaught exception handler that runs the shutdown method if an uncaught exception is thrown by Hydra

If an IOException is caught when trying to start hydra, the shutdown method is called instead of only shutting down the server.