Fixes broken configuration when modifying stages in the GUI, if those stages were not inserted by the GUI
All stage configurations that the GUI touches will now get the fields stageName, stageGroup and libId added to their properties map.
This means the stagegroups endpoint returns a modified properties map that contains the fields, and adding a stage using the Admin Service automatically adds the fields.
The stages endpoint will return only the fields that are currently stored in the database, with no additions.